+++ NEWS.md
@@ -1,6 +1,6 @@
 # enrichplot 1.9.4
 
-+ add `node_label_scale` parameter to adjust the size of node label in `emapplot` function (2020-09-18, Fri)
++ add `node_label_size` parameter to adjust the size of node label in `emapplot` function (2020-09-18, Fri)
 
 
 # enrichplot 1.9.2